+++ src/Stack/Options/GhcBuildParser.hs
@@ -14,7 +14,14 @@ ghcBuildParser hide =
     option
         readGHCBuild
         (long "ghc-build" <> metavar "BUILD" <>
-         completeWith ["standard", "gmp4", "nopie", "tinfo6", "tinfo6-nopie", "ncurses6", "integersimple"] <>
+         completeWith [ "standard"
+                      , "gmp4"
+                      , "nopie"
+                      , "tinfo6"
+                      , "tinfo6-nopie"
+                      , "ncurses6"
+                      , "int-native"
+                      , "integersimple"] <>
          help
              "Specialized GHC build, e.g. 'gmp4' or 'standard' (usually auto-detected)" <>
          hideMods hide